Refactored disk/device handling, now split into two parts and code improved (PXE is temporarily broken).
The device layer is what the rest of the loader sees, manages a list of
devices with a name and a filesystem attached to it. The disk layer is used
internally, mainly by the filesystem (although filesystems don't have to be
on disks). The split was made because there are cases where there is a
filesystem that isn't based on a disk, such as PXE/TFTP, and some code that
will be added later to support a TAR file loaded along with the bootloader
containing the stuff to load.