Changes to the KBoot network boot information.
Added a hardware type and hardware address length field. Also changed the
length of kboot_mac_addr_t to be 16 - PXE's MAC address type is that long,
and although I'm not aware of any MAC address formats that are that long,
it's probably that length in PXE for a reason.