Fixed alignment of heap allocations.
Even though kmalloc() was aligning allocation sizes up to 8 bytes, it still
only ended up with 4 byte aligned allocations because sizeof(heap_chunk_t)
was not a multiple of 8 bytes. This was causing alignment faults on ARMv6
when performing doubleword writes. Fixed by padding out heap_chunk_t by
no longer using the low byte of size as an allocated flag and just storing
it as a separate bool.