Support running from a non-identity mapped virtual address space
Both x86 and ARM run KBoot with the MMU disabled, so everything assumed
that physical addresses can be used directly. On MIPS this is not the
case - the loader runs in the virtual address space from KSEG0. Added a
new architecture definition, LOADER_VIRT_OFFSET, which specifies an
offset to add to a physical address to obtain an accessible virtual
address, as well as P2V and V2P macros to convert physical to virtual
addresses and vice-versa. Also added a LOADER_PHYS_MAX definition which
specifies the highest physical address accessible to the loader, and
changed phys_memory_alloc() to not allocate addresses higher than that.

Signed-off-by: Alex Smith <alex@alex-smith.me.uk>